==> studies/study1/1_study1_data.qmd <==
---
title: "Simulating study 1 data"
author: "Jenna Landy"
---

```{r}
library(MCMCpack)
library(glue)
library(bayesNMF)

data_dir <- "../../data/study1"
```

```{r}
Gs <- c(16, 32, 64, 128)
Ns <- c(2, 4, 8, 16)

reference <- get_cosmic()
K <- nrow(reference)

Nreps <- 10
```

```{r}
# set hyperparameters for distribution of m_g ~ NB(r * N, p)
p <- 0.1
mean_mutations_per_signature <- 1000 # r(1-p)/p
r <- mean_mutations_per_signature * p / (1 - p)

summary(rnbinom(10000, size = r*2, prob = p))
```

```{r}
set.seed(321)
for (N in Ns) {
  for (G in Gs) {
    if (G < 4*N) {next} # skip if G < 4*N
    for (rep in 1:Nreps) {
      # sample a subset of N reference signatures
      sigs <- sample(1:ncol(reference), size = N, replace = FALSE)
      P <- as.matrix(reference[,sigs])

      # sample m_g = E[# mutations for subject g]
      # scales with number of signatures, so all signatures have 
      # approximately equal power
      m <- rnbinom(G, size = r * N, prob = p)

      # sample contributions Eg and mutational counts Mg
      probs <- matrix(nrow = N, ncol = G)
      E <- matrix(nrow = N, ncol = G)
      M <- matrix(nrow = K, ncol = G)
      for (g in 1:G) {
        probs[,g] <- rdirichlet(1, rep(1, N))
        E[,g] <- rmultinom(1, size = m[g], prob = probs[,g])
        M[,g] <- rpois(K, P%*%E[,g])
      }

      # save full dataset
      data <- list(
        M = M,
        P = P,
        E = E,
        probs = probs
      )
      saveRDS(data, file = file.path(
        data_dir, 
        glue("N{N}_G{G}_rep{rep}.rds")
      ))
    }
  }
}
```
